requirements.txt
langchain==0.2.8
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.36.0
faiss-cpu==1.8.0
streamlit-extras==0.4.3
openai==1.35.12
tiktoken==0.7.0